+++ service/dashboards/interface.go
@@ -124,3 +124,25 @@ type LakeviewService interface {
 	// Update dashboard schedule.
 	UpdateSchedule(ctx context.Context, request UpdateScheduleRequest) (*Schedule, error)
 }
+
+// Token-based Lakeview APIs for embedding dashboards in external applications.
+type LakeviewEmbeddedService interface {
+
+	// Read a published dashboard in an embedded ui.
+	//
+	// Get the current published dashboard within an embedded context.
+	GetPublishedDashboardEmbedded(ctx context.Context, request GetPublishedDashboardEmbeddedRequest) error
+}
+
+// Query execution APIs for AI / BI Dashboards
+type QueryExecutionService interface {
+
+	// Cancel the results for the a query for a published, embedded dashboard.
+	CancelPublishedQueryExecution(ctx context.Context, request CancelPublishedQueryExecutionRequest) (*CancelQueryExecutionResponse, error)
+
+	// Execute a query for a published dashboard.
+	ExecutePublishedDashboardQuery(ctx context.Context, request ExecutePublishedDashboardQueryRequest) error
+
+	// Poll the results for the a query for a published, embedded dashboard.
+	PollPublishedQueryStatus(ctx context.Context, request PollPublishedQueryStatusRequest) (*PollQueryStatusResponse, error)
+}