+++ service/serving/interface.go
@@ -63,7 +63,7 @@ type ServingEndpointsService interface {
 
 	// Make external services call using the credentials stored in UC
 	// Connection.
-	HttpRequest(ctx context.Context, request ExternalFunctionRequest) (*ExternalFunctionResponse, error)
+	HttpRequest(ctx context.Context, request ExternalFunctionRequest) (*HttpRequestResponse, error)
 
 	// Get all serving endpoints.
 	//